Added ai_helper.print_eval() function and use
This also includes a time stamp at beginning and end of each CA
evaluation.  It produces _a lot_ of output, mostly meant for figuring
out which evals take too long.
Ugh.  This was a tedious and brainless job, but I'm too tired here on
the airplane for things that take more thought, so this was a good use
of this time.  :-P