Calculate damage for best defense and flat defense
Since when defending you can often force the enemy to be on flat
terrain, use that for defensive calculations.
When attacking, the reverse is true, so assume the opponent has the
best possible defense.
It may be worth doing a village defense option as well, with the
intention of looking for units good at attacking villages.