Fred zone hold: switch to a two-tier rating system
It is not possible to have one rating function that works for hexes
threatened by enemies and unthreatened hexes simultaneously.  Thus, we
separate the two rating systems.
This is only the first step in the process, setting up the mechanism.
Currently, this always chooses the threatened hexes over the
unthreatened ones, which is obviously not right.  That needs to be
sorted out.