Fred: reevaluate after leader-move forced recruiting (bug fix)
When the leader leaves the keep, recruiting is forced first.  This
recruiting changes the situation on the map.  In the worst case, the
hex the leader might want to move to could then be blocked, which
previously lead to an AI crash.  Thus, we now reevaluate the next move
after such recruiting occurs.