Fred: discourage recruiting on hexes needed for planned moves
When the leader is about to move, recruitment needs to happen first.
This recruiting should try to avoid hexes onto which the leader (or
another unit currently planned to move) is trying to get, if possible.
If not possible, recruitment will be done anyway and a new move or move
sequence will be calculated.