Fred: fix bug sometimes making move sequences impossible
If a unit participating in the current move sequence has to move out of
the way for another unit of the same sequence first, it might then not
be able to get to its goal hex any more.  Previously this caused the AI
to crash.  If this happens now, the move sequence is abandoned and
reevaluated.