battle_calcs: revamp the attack rating system
The basic rating system remains the same, but it really needed to be
able to deal with attack combos (which it does now).  For example,
villages or regeneration was counted multiple times in attack
combinations.
This also includes pulling the calculation of att_stats and def_stats
out of get_attack() and a bunch of other related but not separable
changes.