fred_gamestate_utils: add new function get_gamedata()
This combines all the gamestate tables into a single table, so that it
can be passed as a single table to functions needed a lot of this
information.  The individual tables are (for the most part) still
returned by get_gamestate(), as that is slightly faster than always
dissecting the big table.