Fred: refactor attack evaluation to use new utils functions
*** This version should not be used for playing ***

While it is functional, there are quite a few things that are currently
not working very well (or entirely disabled, such as village grabbing),
but this is already a monster commit, so I want to get it out there.
This uses all the new (and faster) utility functions, but it meant that
the entire attack code had to be rewritten.  Other parts of the code
have not been adapted yet.
It also includes some general changes/fixes (such as not caching
counter attacks, as that gives wrong results) that did not make sense
to separate out.