Fred attack utils: rework calc_counter_attack()
Have the function take care of putting units into their hypothetical
positions on the map, rather than the calling functions.  Also, add a
local function to it that adjusts the gamedata tables.
This avoids code duplication and ensures consistency.  It also fixes a
couple bugs where the wrong information was used for attack ratings of
counter attacks.