Fred move_leader_to_keep: wait if best keep hex is occupied
by a unit that can move out of the way, rather than forcing it to move
out of the way (or the leader not getting there and ending the leader’s
move, which is how it was before).  That way, that other unit will not
go in a direction that might keep it from getting to where it should be.