fred_attack_utils.attack_rating(): make attacker_rating negative
All ratings should be defined in a consistent way (larger value means
better) so that they can just be summed up.  Previously,
attacker_rating, which is essentially the damage done to the attacker,
was positive and needed to be subtracted from the overall rating.
That’s unintuitive and can lead to confusion down the road.