Fred: add map analysis layer (huge and ugly commit)
First draft of a complete rework of the Fred AI, adding a map analysis
layer on top of the CAs and zone control actions.
This is a humongous and very dirty commit, but it does not make sense
to separate it into parts.
It is the first working version of the new AI, although with a lot of
very bad code, obsolete code parts, etc. — and it plays very badly at
this point.  But this is the first instance where we’re back to a
(technically) working AI, so I a committing here so that future changes
are logged/checked/controlled more easily.