Fred: base holding score mostly on terrain only, not counter attack outcome
Don’t base it on counter attack outcome any more, but mostly only on
terrain.  Attack outcome is only used for tie breakers and to
disqualify holding combinations.  The attacks are, however, calculated
in order to find the hexes from which the enemy will/can attack and the
units that can get there.
This is done because otherwise the AI will always prefer positions
which more (or fewer, depending on the scoring) enemies can reach.