Fred advancing: add second evaluation of villages not owned by AI
Previously, a unit which can take two villages but has a higher rating
than another unit which can get to only one village might take that one
village away from the latter unit.  This happened in particular in the
NW of the Freelands map.  This commit fixes that problem by giving
higher priority to units which can take fewer villages.