Fred holding: make the pre-rating and the final rating consistent
The final rating was adapted earlier but the pre-rating was left
unchanged (because it identified approximately the same hexes as the
best).  This commit makes both parts use the same type of rating, just
that the pre-rating is a quicker and more approximate version done for
all zone hexes, while the final rating does a more accurate job on a
smaller number of hexes.