Fred events: reorganize how the events are set up and called
This is done so that they (and in particular the events that will be
added next) will only happen for Fred.

For the most part, the Lua code is now only used to set up the events.
It still executes a few things, simply because it is easier to check
certain conditions in Lua. But all of the rest is moved into the WML
code, and is simply set up/triggered from Lua.