Fred attacks: improve attack ratings
The attack combination rating now takes delayed damage (healing,
poison, etc.) into account more accurately.  That is, at the times when
it occurs (at the respective side’s turn), rather than right at the end
of the attack.
The counter attack evaluation still has some big problems (such as only
caring about the worst of the counter attacks on a single unit), but
that’s going to be non-trivial to fix and will be left for later.