Fred recruiting: check that units can in practice move out of way
It’s not enough to check that units have MP left, they also need to
have an empty hex to move to.  This does not apply very often, requires
enemies and AI units blocking the unit in just the right way, but it is
possible (and has happened during AI vs. AI tests).