Fred overall strategy layer: refactor and start to rework parts of it
This is the beginning of a more detailed up-front analysis of how to
distribute units on the map.  It includes things such as:
- making power_missing specific to the action (rather than the zone)
- adding non-immediate threats (threat2)
- prioritization of the zones relative to each other
It’s a huge commit, but it is only the beginning of this work and does,
in fact, contain some regressions.  It is the first point where Fred is
working again (at least technically).