Fred hold utils: first draft of new function get_hold_units()
This is for up front analysis of which units to move to which zone. It
is still very rough, and contains duplicate code, as well as
unnecessary code, w.r.t. the holding code in fred.lua.
Needs to be cleaned up, but committing now so that changes are captures
in the log.