Fred: first rough version of new operations layer, holding and advancing code
The operations layer of the defend_zones stage was entirely redone and
extended to move more of the decision process from the tactical to the
operations level. To go along with this, the holding and advancing code
was adjusted and significantly simplified (and sped up). This is the
first version that produces a working AI, it's a massive commit but
splitting it up into pieces is not possible.

All of the changed parts are only first, incomplete versions, with lots
of missing pieces, but the AI is operational again. In addition, the
attack code has not been adapted yet (but does work) and the leader and
all_map stages are disabled. Thus, current gameplay is not very good
(but still better than the default AI).