Fred holding: another major reworking of the holding code
This one now does the exclusion of hexes due to influence and expected
damage up front. It contains separate (full) ratings for protecting and
"normal" holding. And it choses advancing over holding when there is no
threat to villages and the only holding needed is in the back of the
zone (against enemies coming around the back from other zones).