Fred holding: caluclate damage fraction with respect to current hitpoints
… rather than w.r.t. max_hitpoints, even though the absolute value is
obtained later by multiplying by max_hitpoints.  This is done
intentionally, in order to discourage the use of injured units.

It was already done like this in the two other places in the holding
code where the same type of calculation is done.  I missed this line
when I changed that.