Fred operations layer: new method of assigning units to zones
There are many changes from the previous method, but probably the most
important is that we now allow units to be reassigned from the zone
they are in to another where they are needed more.

This is the first draft of this code. It is likely that additional
tuning is needed.