Fred villages: add unthreatened villages on enemy side of map as goals
Otherwise these might be bypassed as part of the holding moves (since
they are outside the holding zone, and advancing happens after
holding).  They must only be added for direct grabbing, not for moving
units toward them (exploring).

Threatened villages on that side of the map do not need to be included
as they are covered by the holding actions.