Fred: extend all zones to cover enemy keep
Otherwise we get some weird behavior in the endgame, when the AI units
just do not approach the keep.
There is some potential for other behavior being affected, but
everything I have found so far does not matter.  For example, it is not
deterministic now to which zone the village closest to the enemy keep
is assigned.  But that doesn’t really make a difference in practice.