Lua code: use consistent formatting for brackets
Mostly just updating and fixing the formatting.  One difference to
before is that I now sometimes use double brackets without a space for
WML tags for clarity (but not when including a table inside another
table when that is not WML).

Did not change ML recruiter code and debug.lua.