Fred holding: only hold against leader threat enemies when protecting leader
As opposed to all enemies in the zone.  This is needed to get the
between_map right.  It also includes an update on the threshold for
this map.

I’m not clear on whether this is the best thing to do.  Probably a mix
of the two is needed eventually.