Fred attack utils: ensure counter attackers have HP>0
This is only relevant for counter attacks, not forward attacks, as the
unit stats are changed for the former.

Interesting note: I found this because the sum of the attack outcome
statistics did not add up to 1 in one case with a 0 HP attacker.  In
some rare cases (I think it only happens when both attacker and
defender have exactly one strike), the statistics returned by
wesnoth.simulate_combat are incomplete then.