Fred attack combo eval: pass best weapon to exec function
Previously, the weapon choice in the attack execution was left to the
Wesnoth engine. This could mean that the weapon used in the evaluation
was different from that in the execution in some cases. For example, an
Orcish Slayer leader might end up attacking a grunt with the melee
attack rather than the ranged attack because the CTK is higher, but get
itself into a precarious situation that way.

As a downside, we now always use max_damage_weapon, which makes the
option kind of useless and is also not always the best choice. This
should be improved at a later time.