Fred operations: add max_taken_any_weapon to unit_attacks for counter attack
The other damages in the table are for the “best” weapon of the enemy
in the counter attack, that is, also taking the AI unit’s weapon into
account. However, for some applications we simply want the highest
damage weapon, independent of ramifications.

Not: the max_damage field in unit_info does not take resistances into
account.