Fred attack utils attack_combo_eval: remove ctd_limit
This was in place so that enemy units close to dying did not get
overrated, making an otherwise unfavorable attack look good. The
problem with it was that it introduced a discontinuity, which could
lead to the worse attack being selected if its CTD for the enemy was
just below the limit, while being just above the limit for the better
attack.

Removing this means that this effect needs to be taken into account
otherwise, ideally using smoothly varying functions.  For now, this is
left to the calling functions and not done in attack_combo_eval.