Fred holding: prefer the highest-HP units
This is the exact opposite of how it was before. The previous
motivation was that the weakest unit should get the best hex in order
to protect it. However, putting the strongest unit on the best hex will
make it much harder for the enemy to remove it from there.
This might have to be refined a little, esp. when it comes to villages.