Fred ops: keep already protect hexes in protect_locs
Add a flag to mark them as protected. There’s a small change in
behavior going along with this as previously other (farther back) hexes
were added to protect_locs if unprotected. That is not the case any
more. As we need to enable multiple protect_locs soon anyway, that
doesn’t really matter.

There’s also a small chance to advancing units that are not currently
in the zone, but if anything, this should be for the better.