Fred ops: completely rework leader objectives eval
Note that this results in Fred not being entirely operational, as this is only the evaluation of the objectives, with the evaluation of actions not done yet.  This is done because I want the commits to be (somewhat) separated and separable, so that it is (somewhat) easier to see later what was done.