Fred ops: combine leader, village and unit protection eval
This is only the operations evaluation of what to protect so far.  It is not taken into account by any of the actions yet.  The code is also distributed over different functions in different files at the moment.  This needs to be cleaned up once all the evals and functionality are there.