Fred ops: shorten how units are stored in tables
They don't need to be stored in id = loc format, as the location is already directly accessible from move_data.units.  For now, this mostly shortens how the tables are displayed in debug mode, which makes them easier to read.  This does not cover all uses yet, mostly just those in the operations analysis.