Fred ops/exec: add used_units table
Done only on unit moves, and only if units do not have moves left after. Recruits are excluded (as they do not have an assigned zone) and attacks are not considered because there should be a move for those units beforehand, either on the same turn or a previous turn.