Fred ops: only calculate enemy_leader_distance_maps when needed
That is, when it does not exist in ops_data (either at beginning of a turn, or when it was reset because the zones changed). The enemy_leader_distance_maps calculation is expensive, but does not depend on the own leader location or leader goal, so for the most part it only needs to be done once per turn.