Fred attacks: switch to positve vs. negative rating system
As opposed to the attacker vs. defender system that was used previously. The problem with the old system was that there could be both positive and negative contributions to both the attacker and defender, and thus the weighting using value_ratio did not always work as intended. With the new system, value_ratio does not quite weight damage received vs. damage done any more, but it should overall be better.

This includes re-(mis-)naming some table fields to start with yyy_. This is done for testing purposes, to make sure none of the old ratings are accidentally used somewhere. It will be cleaned up before the next release.