Fred advancing: redo evaluation based on ops advancing goals
This is now based on the ops advancing goals, rather than the unstable switch between toward own and enemy leader. The discontinuity between threatened and unthreatened hexes can be removed with that too.

Note that this is done in a way to emphasize the changes in the diff.  The code will be cleaned up in the next commit.