Fred ops: do not hold if other units are already holding
If other units are already holding in the zone and there is nothing to protect (incl. those other units), then don't hold there.  This is meant to prevent new units from moving in front of those already holding.

The current implementation of what units are already holding and which of them need to be protected is too simple, but this is a start of putting the mechanics in place.