Fred holding: restrict list of protect hexes
Previously they had to be adjacent to hexes the enemy units can reach. Now they have to be adjacent to those the enemy can reach AND from there get next to the hexes to be protected.
This may be a bit overly restrictive, but it is better than the previous version.