Fred advancing: rework advance goals assignements
If units are holding already in the zone, try to stay close to them.  If not, holds were not successfully found, so we should be careful with advancers as well and try to stay away from enemies.

The latter assumes that we always try to hold, which is currently not the case, so that also needs to be changed.