Fred get_between_map: add weights for 'locs'
In addition to the already existing weights for units. These are based on the exposures of the different goals.

This does not quite work in all situations, but the full commit will shuffle things around so much that it will be impossible to see what the actual substantial changes are. So I am going to split this up into several commits.