Fred: check whether units can move using my_units_noMP
This is as opposed to checking whether they have moves left.  The reason is that a surrounded unit might have moves left but is not able to move anyway.

In many cases this makes no difference, but sometimes an action (such as a more risky attack) is/should be taken by a unit that cannot move, that would not be taken if it could move away.