Fred ops: add nostrikeback table to unit_attacks
This contains information about the best attack by a unit at a range at which the enemy does not have a weapon. This can be used to determine whether an attack with a weapon that is not the main weapon should be done, when the attack with the main weapon is too risky.