Fred: rename zones west/east -> right/left
This is done so that they can be defined with the same names for Side 2.  Note that this is right/left in the direction in which the AI units are moving, which means (for Side 1) it is opposite to the sides on which they appear on the map to the player.