Fred ops: set up mechanism for actions with multiple scores
This allows an action evaluation to return as score inside the 'action' table which is different from (lower than) the config rating. If this is done, the next action(s) are checked first and executed, if they have a higher score. Otherwise the previously evaluated action is executed.

This commit also includes cleanup of the debug output, which greatly facilitates following what is going on.